requirements.txt
aiogram==2.25.2
python-dotenv==1.0.0
aiosqlite
asyncpg
yookassa==3.3.0
